fix damage adjacent to
