+++ Whenever.DmgTypeEtcExt.Experimental/Effects/DamageAdjacentToCombatantEffect.cs.meta
