+++ Whenever.DmgTypeEtcExt.Experimental/Effects/DamageCombatantEffect.cs.meta
